write some initial tests
